crates/sel4-static-heap: Remove
