+++ Cargo.toml
@@ -137,7 +137,6 @@ members = [
     "crates/sel4-shared-ring-buffer/bookkeeping",
     "crates/sel4-shared-ring-buffer/smoltcp",
     "crates/sel4-stack",
-    "crates/sel4-static-heap",
     "crates/sel4-sync",
     "crates/sel4-sync/trivial",
     "crates/sel4-synthetic-elf",
