--- crates/sel4-static-heap/Cargo.nix
@@ -1,11 +1,0 @@
-#
-# Copyright 2024, Colias Group, LLC
-#
-# SPDX-License-Identifier: BSD-2-Clause
-#
-
-{ mk }:
-
-mk {
-  package.name = "sel4-static-heap";
-}
